fix : build issue fixed

+++ app/api/servers/[serverId]/leave/route.ts
@@ -1,6 +1,7 @@
+import { NextResponse } from "next/server";
+
 import { currentProfile } from "@/lib/current-profile";
 import { db } from "@/lib/db";
-import { NextResponse } from "next/server";
 
 export async function PATCH(
   req: Request,
@@ -13,29 +14,34 @@ export async function PATCH(
       return new NextResponse("Unauthorized", { status: 401 });
     }
 
+    if (!params.serverId) {
+      return new NextResponse("Server ID missing", { status: 400 });
+    }
+
     const server = await db.server.update({
       where: {
         id: params.serverId,
         profileId: {
-          not: profile.id,
+          not: profile.id
         },
         members: {
           some: {
-            profileId: profile.id,
-          },
-        },
+            profileId: profile.id
+          }
+        }
       },
       data: {
         members: {
           deleteMany: {
-            profileId: profile.id,
-          },
-        },
-      },
+            profileId: profile.id
+          }
+        }
+      }
     });
+
     return NextResponse.json(server);
   } catch (error) {
-    console.log("[SERVER_LEAVE_PATCH]", error);
+    console.log("[SERVER_ID_LEAVE]", error);
     return new NextResponse("Internal Error", { status: 500 });
   }
 }
